requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.9.1

faiss-cpu==1.8.0
sentence-transformers==2.2.2
numpy==1.26.4
regex==2024.5.15

requests==2.32.3